# protect the executor against reentrant tasks