@glommer 2021-10-03
I must be missing something. In terms of results themselves, while you claim "A gets better while nothing gets worse", a lot of benchmarks are actually worse on "After". The cost of a contended semaphore is higher, the cost of running a task in a task queue is higher, and even the ones that do complete on the noise benchmark are sometimes higher.

I don't fully understand how you define a "reentrant task", nor why this concept is needed or related to the problem you describe. There are places where we manually adjust the vruntime of a queue. For example, when the task queue has been idle. Why can't you just adjust the vruntime on explicit yields ?

Also, it may be worth checking your application. While we can obviously improve things, maybe the solution involves stopping using them. `yield_if_needed()` is a much better pattern. I understand avoiding using it because there are known issues with how fast Rust's TLS is, but if that creates another problem of a different kind then all of a sudden TLS access doesn't sound that expensive in comparison.